refactor(channel): fold daemon/telegram.rs into channel/telegram.rs (T1c)
Moves `notify_telegram` / `notify_telegram_silent` out of
`src/daemon/telegram.rs` (deleted) into `src/channel/telegram.rs` so all
Telegram Bot-API code lives behind a single module. Follows the
single-file-per-channel convention established by T1b.

Consumer updates:
- `src/daemon/mod.rs`: drop `mod telegram;`, switch
  `use telegram::notify_telegram` → `use crate::channel::telegram::notify_telegram`.
- `src/daemon/supervisor.rs`: switch `use super::telegram::{…}` →
  `use crate::channel::telegram::{…}`.

No behavior change — the moved function body is byte-identical except for
scoping the `Requester` / `SendMessageSetters` imports inside the async
block (mirrors the existing `send_with_topic` helper's pattern) and
calling `handle_send_failure` without the `crate::channel::telegram::`
prefix now that it's a sibling item.

Second of 3 commits in the T1 atomic cut-over PR — do not merge in isolation.
