+++ src/channel/telegram.rs
@@ -775,6 +775,95 @@ pub fn try_download_attachment(instance_name: &str, file_id: &str) -> anyhow::Re
 }
 
 // ---------------------------------------------------------------------------
+// Supervisor / daemon notify helpers (folded from src/daemon/telegram.rs
+// during T1c — single-file-per-channel convention).
+// ---------------------------------------------------------------------------
+
+/// Send a notification to Telegram (instance topic or general).
+pub fn notify_telegram(home: &std::path::Path, instance_name: &str, text: &str) {
+    notify_telegram_inner(home, instance_name, text, false);
+}
+
+/// Send a notification with Telegram's `disable_notification` flag set — the
+/// message still appears in the topic but does not push/vibrate the operator.
+/// Use for state-recovery pings that should not compete with real alerts.
+pub fn notify_telegram_silent(home: &std::path::Path, instance_name: &str, text: &str) {
+    notify_telegram_inner(home, instance_name, text, true);
+}
+
+fn notify_telegram_inner(
+    home: &std::path::Path,
+    instance_name: &str,
+    text: &str,
+    disable_notification: bool,
+) {
+    let config = match crate::fleet::FleetConfig::load(&home.join("fleet.yaml")) {
+        Ok(c) => c,
+        Err(_) => return,
+    };
+    let (token, group_id, topic_id) = match &config.channel {
+        Some(crate::fleet::ChannelConfig::Telegram {
+            bot_token_env,
+            group_id,
+            ..
+        }) => match std::env::var(bot_token_env) {
+            Ok(t) => (
+                t,
+                *group_id,
+                config.instances.get(instance_name).and_then(|i| i.topic_id),
+            ),
+            Err(_) => return,
+        },
+        None => return,
+    };
+
+    let text = text.to_string();
+    let home_owned = home.to_path_buf();
+    let instance_owned = instance_name.to_string();
+    std::thread::Builder::new()
+        .name("tg_notify".into())
+        .spawn(move || {
+            let Ok(rt) = tokio::runtime::Builder::new_current_thread()
+                .enable_all()
+                .build()
+            else {
+                return;
+            };
+            if let Err(e) = rt.block_on(async {
+                use teloxide::payloads::SendMessageSetters;
+                use teloxide::prelude::Requester;
+                let bot = teloxide::Bot::new(&token);
+                let chat_id = teloxide::types::ChatId(group_id);
+                match topic_id {
+                    Some(tid) if tid != 1 => {
+                        let mut req = bot.send_message(chat_id, &text).message_thread_id(
+                            teloxide::types::ThreadId(teloxide::types::MessageId(tid)),
+                        );
+                        if disable_notification {
+                            req = req.disable_notification(true);
+                        }
+                        req.await?;
+                    }
+                    _ => {
+                        let mut req = bot.send_message(chat_id, &text);
+                        if disable_notification {
+                            req = req.disable_notification(true);
+                        }
+                        req.await?;
+                    }
+                }
+                Ok::<(), anyhow::Error>(())
+            }) {
+                let handled = handle_send_failure(&e, &home_owned, &instance_owned, topic_id, None);
+                if !handled {
+                    tracing::warn!(error = %e, "telegram notify failed");
+                }
+            }
+        })
+        .ok();
+}
+
+// ---------------------------------------------------------------------------
 // Channel trait adapter (T1b scaffold)
 //
 // `TelegramChannel` wraps the legacy `Arc<Mutex<TelegramState>>` so external
